Update angular-csv-import.js
Wraped strictly executed code into a imidiately executed closure as it is commonly done in todays javascript libraries. This prefents issues from happening if the framework will be used in a probjekt in combinations with a bundleing framework and other javascript libararies that aren't marked as strict. Had an issue with that resulting in 'use strict' affecting the entire bundled script file which leaded to trouble when using firefox.